requirements.txt
# Core Dependencies
alpaca-trade-api>=3.0.0
yfinance>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
ta-lib>=0.4.0
scikit-learn>=1.3.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0

# Financial Data
yfinance>=0.2.0
pandas-datareader>=0.10.0

# Technical Analysis
ta>=0.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# News & Sentiment
newspaper3k>=0.2.8
google-search-results>=2.4.0

# API & Web
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Development
pytest>=7.4.0
black>=23.0.0